integration_tests/pyproject.toml
[tool.poetry]
name = "pyheap-integration-tests"
version = "0.1.0"
description = ""
authors = ["Ivan Yurchenko <ivan@ivanyu.me>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = "^22.6"
dateutils = "^0.6.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
